Moved body reader tests to body package.
